aider: Remove the part of the `render_gui` function that prints the messages in the bottom right part of the screen.
# Aider chat conversation:

USER: Yes, we commented out the correct one. We can now remove it.

ASSISTANT: To remove the message log that appears in the bottom right part of the screen, we need to modify the `main.py` file. Specifically, we need to remove the part in the `render_gui` function that prints the messages.

Here is the *edit block* for this change:

```python
main.py
<<<<<<< HEAD
    # y = console.height - message_console_height - 2
    # if isinstance(message_log, MessageLog):
    #     for message in message_log.messages:
    #         console.print(console.width - 2, y, str(message), fg=(255, 255, 255), bg=(0, 0, 0), alignment=tcod.RIGHT)
    #         y -= 1
    # else:
    #     raise TypeError("message_log must be an instance of MessageLog.")
=======
>>>>>>> updated
```

This change removes the part of the `render_gui` function that prints the messages in the bottom right part of the screen. The messages should now only appear in the middle bottom part of the screen.