aider: Created Inventory class with methods to add and remove items.
# Aider chat conversation:

USER: We'll also need to make an invetory.py in the 'components' directory

ASSISTANT: Sure, we can create a new file named `inventory.py` in the `components` directory. This file will contain the `Inventory` class which will be used to manage the items that the player has in their inventory.

Here is the *edit block* for this change:

```python
components/inventory.py
<<<<<<< HEAD
=======
class Inventory:
    def __init__(self, capacity: int):
        self.capacity = capacity
        self.items = []

    def add_item(self, item):
        if len(self.items) >= self.capacity:
            raise Exception("Inventory is full.")
        self.items.append(item)

    def remove_item(self, item):
        self.items.remove(item)
>>>>>>> updated
```

This change creates a new `Inventory` class with methods to add and remove items. The `add_item` method raises an exception if the inventory is full.