aider: Fixed circular imports in engine/input_handlers.
# Aider chat conversation:

USER: we need to fix the circular imports in engine/input_handlers

ASSISTANT: To fix the circular imports in `engine.py` and `input_handlers.py`, we need to modify the way we are importing the `Engine` and `EventHandler` classes.

We can use a technique called "forward declaration" where we declare the type of the variable in a string. This way, Python won't try to import the module immediately and the circular import can be avoided.

Here are the changes we need to make:

1. In `engine.py`, change the import statement for `EventHandler` from `from input_handlers import EventHandler` to `EventHandler = 'input_handlers.EventHandler'`.

2. In `input_handlers.py`, change the import statement for `Engine` from `from engine import Engine` to `Engine = 'engine.Engine'`.

Here are the *edit blocks* for these changes:

```python
engine.py
<<<<<<< HEAD
from input_handlers import EventHandler
=======
EventHandler = 'input_handlers.EventHandler'
>>>>>>> updated
```

```python
input_handlers.py
<<<<<<< HEAD
from engine import Engine
=======
Engine = 'engine.Engine'
>>>>>>> updated
```

Please confirm these changes.