aider: Fixed issue with placing items in rooms by adding a maximum number of attempts to place an item.
# Aider chat conversation:

USER: I added these *read-write* files: procgen.py

ASSISTANT: Great, now that we have the `procgen.py` file, we can check if the `generate_dungeon` function correctly uses the `max_items_per_room` parameter.

Looking at the `generate_dungeon` function, it seems to correctly use the `max_items_per_room` parameter when calling the `place_entities` function. This function generates a random number of items up to `max_items_per_room` and places them in the room.

However, there might be an issue if the `place_entities` function tries to place more items than there are available spaces in the room. This could happen if the room is small or if there are many monsters in the room. To fix this, we could modify the `place_entities` function to stop placing items when there are no more available spaces.

Here is the proposed change:

```python
procgen.py
<<<<<<< HEAD
    for i in range(number_of_items):
        x = random.randint(room.x1 + 1, room.x2 - 1)
        y = random.randint(room.y1 + 1, room.y2 - 1)

        if not any(entity.x == x and entity.y == y for entity in dungeon.entities):
            entity_factories.health_potion.spawn(dungeon, x, y)
=======
    for _ in range(number_of_items):
        for _ in range(100):  # Maximum number of attempts to place an item
            x = random.randint(room.x1 + 1, room.x2 - 1)
            y = random.randint(room.y1 + 1, room.y2 - 1)

            if not any(entity.x == x and entity.y == y for entity in dungeon.entities):
                entity_factories.health_potion.spawn(dungeon, x, y)
                break  # Item placed successfully
>>>>>>> updated
```

This change will attempt to place an item up to 100 times before giving up. This should be enough attempts for most rooms, but you can increase this number if you have very large rooms or a high number of items per room.